requirements.txt
aiohttp==1.0.5
async-timeout==2.0.0
beautifulsoup4==4.6.0
certifi==2017.7.27.1
chardet==3.0.4
discord.py==0.16.12
idna==2.6
multidict==3.2.0
oauthlib==2.0.4
pendulum==1.3.0
python-dateutil==2.6.1
pytz==2017.2
pytzdata==2017.2.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
tzlocal==1.4
urllib3==1.22
websockets==3.4
